+++ Realm/ObjectServerTests/SwiftSyncTestCase.swift
@@ -242,9 +242,9 @@ open class SwiftSyncTestCase: RLMSyncTestCase {
         let subscriptions = realm.subscriptions
         XCTAssertNotNil(subscriptions)
         let ex = expectation(description: "state change complete")
-        subscriptions.write({
-            subscriptions.append(QuerySubscription<SwiftPerson>(where: "TRUEPREDICATE"))
-            subscriptions.append(QuerySubscription<SwiftTypesSyncObject>(where: "TRUEPREDICATE"))
+        subscriptions.update({
+            subscriptions.append(QuerySubscription<SwiftPerson>())
+            subscriptions.append(QuerySubscription<SwiftTypesSyncObject>())
         }, onComplete: { error in
             XCTAssertNil(error)
             ex.fulfill()